requirements.txt

fastapi
uvicorn
pandas
numpy
ta
requests
python-binance
